Replaced iconSvgPath with icon titles
