+++ src/components/common/Icon.tsx
@@ -8,5 +8,9 @@ interface IconProps extends icons.IconProps {
 
 export const Icon = ({ iconName, ...props }: IconProps) => {
   const BootstrapIcon = icons[iconName]
+  // set title prop to iconName if not set
+  if (!props.title) {
+    props.title = iconName
+  }
   return <BootstrapIcon {...props} />
 }
