+++ src/constants/icons.ts
@@ -1,0 +1,55 @@
+import { IconName } from '@/components/common'
+
+// List of icons to display in the icon picker
+export const iconPickerIcons: IconName[] = [
+  'ChatFill',
+  'PencilFill',
+  'Truck',
+  'TelephoneFill',
+  'ChevronLeft',
+  'ChevronRight',
+  'ArrowUp',
+  'ArrowDown',
+  'ArrowClockwise',
+  'EnvelopeFill',
+  'FolderFill',
+  'VolumeUpFill',
+  'CurrencyEuro',
+  'CurrencyDollar',
+  'SunFill',
+  'Square',
+  'SquareFill',
+  'FlagFill',
+  'EyeFill',
+  'PrinterFill',
+  'ChatDotsFill',
+  'MapFill',
+  'CheckCircle',
+  'Circle',
+  'CircleFill',
+  'Hexagon',
+  'HexagonFill',
+  'PuzzleFill',
+  'StarFill',
+  'Film',
+  'CaretUp',
+  'CaretDown',
+  'CaretLeft',
+  'CaretRight',
+  'CloudFill',
+  'CloudDownloadFill',
+  'CloudUploadFill',
+  'Code',
+  'ClipboardFill',
+  'CollectionFill',
+  'GearFill',
+  'Heart',
+  'HouseFill',
+  'BalloonFill',
+  'AirplaneFill',
+  'SuitClubFill',
+  'SuitDiamondFill',
+  'SuitHeartFill',
+  'SuitSpadeFill',
+  'CalendarDateFill'
+]
